crm-service/requirements.txt
kafka-python>=2.0.0
boto3>=1.26.0
